+++ src/program/protocols/union/types.rs
@@ -5,7 +5,6 @@ use bitcoin::{PublicKey, Txid};
 use key_manager::musig2::{secp::MaybeScalar, PubNonce};
 use protocol_builder::types::OutputType;
 use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
 use uuid::Uuid;
 
 use crate::{
@@ -50,7 +49,6 @@ pub const FUNDS_ADVANCED: &str = "funds_advanced";
 pub const FUNDS_ADVANCE_SPV: &str = "funds_advance_spv";
 pub const UNION_SPV_NOTIFICATION: &str = "union_spv_notification";
 pub const FULL_PENALIZATION_DATA: &str = "full_penalization_data";
-pub const UNION_SETTINGS: &str = "union_settings";
 
 // Transaction names
 pub const REQUEST_PEGIN_TX: &str = "REQUEST_PEGIN_TX";
@@ -107,8 +105,6 @@ pub const WT_DISABLER_DIRECTORY_UTXO: &str = "WT_DISABLER_DIRECTORY_UTXO";
 pub const OPERATOR: &str = "OP";
 pub const WATCHTOWER: &str = "WT";
 
-pub const GLOBAL_SETTINGS_UUID: Uuid = Uuid::from_bytes(*b"UNION_BRIDGE-000");
-
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct MemberData {
     pub role: ParticipantRole,
@@ -123,9 +119,7 @@ pub struct Committee {
     pub dispute_aggregated_key: PublicKey,
     pub packet_size: u32,
     pub stream_denomination: u64,
-    pub pegin_confirmations: u32,
-    pub pegout_confirmations: u32,
-    pub reject_pegin_confirmations: u32,
+    pub settings: PacketSettings,
 }
 
 impl Committee {
@@ -332,7 +326,10 @@ impl FullPenalizationData {
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct StreamSettings {
+pub struct PacketSettings {
+    pub pegin_confirmations: u32,
+    pub pegout_confirmations: u32,
+    pub reject_pegin_confirmations: u32,
     pub short_timelock: u16,
     pub long_timelock: u16,
     pub op_won_timelock: u16,
@@ -343,14 +340,21 @@ pub struct StreamSettings {
     pub request_pegin_timelock: u16,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct UnionSettings {
-    pub settings: HashMap<u64, StreamSettings>,
-}
-
-impl UnionSettings {
-    pub fn name() -> String {
-        UNION_SETTINGS.to_string()
+impl Default for PacketSettings {
+    fn default() -> Self {
+        Self {
+            pegin_confirmations: 6,
+            pegout_confirmations: 6,
+            reject_pegin_confirmations: 1,
+            short_timelock: 6,
+            long_timelock: 12,
+            op_won_timelock: 250,
+            claim_gate_timelock: 6,
+            input_not_revealed_timelock: 14,
+            op_no_cosign_timelock: 12,
+            wt_no_challenge_timelock: 12,
+            request_pegin_timelock: 12,
+        }
     }
 }